components/dump-service-dev/dump-service-dev/pyproject.toml
[project]
name = "dump-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=3.0.0",
    "aioredis>=2.0.1",
    "alembic>=1.17.2",
    "alembic-postgresql-enum>=1.8.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "boto3>=1.42.5",
    "dill>=0.4.0",
    "fastapi>=0.124.4",
    "greenlet>=3.3.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "pyarrow>=17.0.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "tomlkit>=0.13.3",
    "uvicorn>=0.38.0",
    "ujson>=5.11.0",
    "uuid7>=0.1.0",
    "paho-mqtt>=2.1.0",
    "humanfriendly>=10.0",
]




[tool.ruff]
line-length = 100
extend-exclude = [
    "src/migrations",
]
force-exclude = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # isort
    "I",
    # tidy imports
    "TID",
    # flake8-bugbear
    "B",
    # flake8-bandit
    "S",
    # pydocstyle
    "D",
    # pyugrade
    "UP"
]
ignore = ["D104", "D107", "B008", "E721"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "D100", "D101", "D102", "D103", "F403", "D106"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100
max-line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.0",
    "ruff>=0.11.13,<=0.12.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.8",
    "testcontainers>=4.13.3",
    "pytest-dotenv>=0.5.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
